+++ vpype_cli/operations.py
@@ -379,7 +379,12 @@ def snap(line_collection: vp.LineCollection, pitch: float) -> vp.LineCollection:
 @cli.command(group="Operations")
 @vp.layer_processor
 def reverse(line_collection: vp.LineCollection) -> vp.LineCollection:
-    """Reverse order of lines."""
+    """Reverse order of lines.
+
+    Reverse the order of lines within their respective layers. Individual lines are not
+    modified (in particular, their trajectory is not inverted). Only the order in which they
+    are drawn is reversed.
+    """
 
     line_collection.reverse()
     return line_collection